# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Credit Service
  description: API to manage Bank Credit.
  version: 1.0.0
servers:
  - url: http://localhost:8084
    description: Local server

paths:
  /v1/credit:
    get:
      summary: Get all credits
      operationId: getAllCredits
      responses:
        '200':
          description: List of credits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new Credit
      operationId: createCredit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequest'
      responses:
        '201':
          description: Credit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/credit/{creditId}:
    get:
      summary: Get an credit by ID
      operationId: getCreditById
      parameters:
        - name: creditId
          in: path
          required: true
          description: Id of credit
          schema:
            type: string
      responses:
        '200':
          description: Credit found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update an credit by ID
      operationId: updateCredit
      parameters:
        - name: creditId
          in: path
          required: true
          description: Id of credit
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequest'
      responses:
        '200':
          description: Credit updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete an credit by ID
      operationId: deleteCredit
      parameters:
        - name: creditId
          in: path
          required: true
          description: Id of credit
          schema:
            type: string
      responses:
        '204':
          description: Credit deleted
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/credit/{creditId}/payment:
    post:
      summary: Make a payment from an credit
      operationId: makePayment
      parameters:
        - name: creditId
          in: path
          required: true
          description: Id of credit
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/paymentRequest'
      responses:
        '200':
          description: Withdrawal made
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/credit/{creditId}/charge:
    post:
      summary: Make a payment from an credit
      operationId: chargeCreditCard
      parameters:
        - name: creditId
          in: path
          required: true
          description: Id of credit
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chargeRequest'
      responses:
        '200':
          description: Charge made
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/credit/{creditId}/balances:
    get:
      summary: Get client balances
      operationId: getClientBalances
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Client balances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/credit/{creditId}/transactions:
    get:
      summary: Get client transactions
      operationId: getClientTransactions
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    CreditRequest:
      type: object
      properties:
        type:
          type: string
          enum: [PERSONAL,BUSINESS]
        creditLimit:
          type: number
        interestRate:
          type: number
        clientId:
          type: string
      required:
        - type
        - clientId
        - interestRate
        - creditLimit
    CreditResponse:
      type: object
      properties:
        _id:
          type: string
        creditNumber:
          type: string
        type:
          type: string
          enum: [ PERSONAL,BUSINESS]
        creditLimit:
          type: decimal
        consumptionAmount:
          type: decimal
        balance:
          type: decimal
        interestRate:
          type: decimal
        clientId:
          type: string
        createdDate:
          type: string
          format: date-time
    paymentRequest:
      type: object
      properties:
        amount:
          type: decimal
    chargeRequest:
      type: object
      properties:
        amount:
          type: decimal
    TransactionRequest:
      type: object
      required:
        - productId
        - clientId
        - amount
      properties:
        productId:
          type: string
        type:
          type: string
          enum: [ PAYMENT,CHARGE ]
        amount:
          type: number
          format: double
        clientId:
          type: string
        balance:
          type: number
          format: double
    TransactionResponse:
      type: object
      required:
        - clientId
        - productId
      properties:
        productId:
          type: string
        type:
          type: string
          enum: [ PAYMENT,CHARGE ]
        amount:
          type: number
        clientId:
          type: string
    BalanceResponse:
      type: object
      properties:
        clientId:
          type: string
        balance:
          type: number
    ErrorResponse:
      type: object
      required:
        - status
        - code
        - message
        - timestamp
      properties:
        status:
          type: int
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time